Move download logic
